CPM-257: Split code input
